Level Editor Camera
Level Editor Camera all done, just drag this script onto the level
editor camera and boom all done.

v1.0: Added maxX and maxZ variables for when we know the dimensions of a
dungeon, and can restrain the camera to those bounds. Until then, you
can go in the x or z direction forever.