Bug Fixes!
v1.3: Bug fixes

BUGS SQUASHED: Click and drag movement not moving directly left or
right, but rather along the x and z axis equally.

KNOWN BUGS: After moving with the arrow keys and attempting to click and
drag again causes a reset to the last point click and drag was used.