Drag and Drop Bug Fixes
CameraAdjuster.cs fixes.

BUGS SQUASHED: No longer jumps to last click and drag location. Should
now drag fine, regardless of when the arrow keys are used to move the
camera.

BUGS KNOWN: Possible bug: when binding the y axis of the camera to the
min or max, causes an effect similar to zooming along the z axis. Could
be what we want, could be bug.