Victory, Reward, Quit, and Chicken Out Tested
Fixed problem with character persistence from scene to scene. The
characters now stick around after the selection screen. Quitting will
take the game to the title screen and delete the GSM.

These need to be added to the player class for chicken and victory to
work as intended.

bool atEnd
bool atBegin

These variables must  also be changed in the appropriate functions
where there are placeholders now.